pyproject.toml
[tool.poetry]
name = "honto-sale-to-kindle"
version = "0.1.0"
description = ""
authors = ["hito_horobe"]

[tool.poetry.dependencies]
python = "3.9.5"

[tool.poetry.dev-dependencies]
requests = "^2.28.1"
python-amazon-paapi = "^4.3.0"
bs4 = "^0.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
